Refactor map views with MapViewFragment
